refactor: ATLAS-43: Changed test namespaces to be in line with units under test.
